Fixed major rendering bugs and crashes
there's still rounding inaccuracies when it comes to zooming in,
which i'm going to try to iron out. rendering is also a lot more
efficient now because the amount of if statements in the rendering
loop was decreased to zero. the loop will only run for the tiles that
are visible on screen, rather than going through the whole thing and checking
as it's looping. so yea that stuff