Input system overhaul
draggin the board around is now possible by clicking and dragging with
the mouse, and it's a lot smoother than before. there's a few quirks that
still need to be ironed out, but overall it's a pleasant experience.

up next: getting rid of that stupid camera jitter when far off from the
central origin of a board